Updated Settings and Configs for Symfony 4 Flex
I have updated the composer.json requirements to allow for Symfony 4 as well I have taken out the requirement for symfony/symfony since flex conflicts with that repository. Now composer is requiring symfony/framework-bundle allowing all the symfony component to be required or not but to as well load flex.

This files update reflects the setup/config files that you can utilize. When I updated via composer flex did not auto add the bundle to my bundles.php file, or add a flex recipe. I had to manually add these, but the setup notes are above. :D hope this helps.